requirements.txt
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
playwright>=1.40.0
keyring>=24.0.0
